Split v2 eval comparison helpers
